backend/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
sqlalchemy>=2.0.0
alembic>=1.12.0
psycopg2-binary>=2.9.9
pgvector>=0.2.4
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-dotenv>=1.0.0
cryptography>=41.0.0
sentence-transformers>=2.2.2
dashscope>=1.14.0
httpx>=0.25.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
bcrypt==3.2.2
python-multipart>=0.0.6
PyMySQL>=1.1.0

# 文件解析
pdfplumber>=0.10.0
python-docx>=1.1.0
python-pptx>=0.6.23

# 工具箱：PDF 转 Word（高质量，接近 WPS 效果）
pdf2docx>=0.5.8
